aws-ls-s3/requirements.txt
boto3
pyyaml
awscli
